I just improved the sorting function a little bit
I added a part where the program would accept an order argument which defaults to ascending order to make the sorting to your liking.

It's kinda a neeat feature I dunno. Either way, I also made changes to the loop shits, especially the loop condition of the bse main loop. I then changed the assignment of the unsorted arr and the sorted in ehich the unsorted reference array would not be passed by reference to the sorted variable. This results to isolated events among the changes of the swapElements() function.