I just finished the end panel shits at fucking last
I tweaked alot of the existing functions just to make the the end panel compatible especially on going going to the previous panel, the user form

Thank god that I really made the codebase kinda modular and not repeatitive. If there's lines of code that would repeat I would slap them in a function for fuck sakes.