Added html and css design for the start and end panels. And then also, before I implemented those shits, I finished first the submission functionality. So yeah, that's all I guess.
Also I did some minor tweaks on the responsive breakpoints. Made it more shorter and flexible so it kinda works more elegant in a way --I mean the resizing part, you know

Just so you know, I just did the startend panels in a whim and forgot about git in a moment. I should've fucking commited first and started a seperate branch for the startend panel but I was fucking forgetful. And when I realized that, I said "Ehhhh, nagawa ko na pare", so I just proceeded further on with no commits and push in a span passing over 28 hours, I guess?