All functionalities DONE!
I've done it at fucking last!

Added the succes panel shit, discovered and fix some bugs along the way, those kinds of shit, and then that's it I guess

My only uncheck thing is to edit the admin panel for my cousin, my groupmate, to fucking understand the shit very clearly. I just need it to be intuitive for her. I'm still not sure whether to make a new page for the analyzing of all the answer or just edit the admin panel like what I said earlier, but still, my main priority now is to like make the fucking shit for my cousin to be easy and intuitive for her. I just don't want her to be felt dragged. I mean this was my idea to say the least, so that's like one of my obligations to say the least.