Fixed the sorting function issue at last!
Turns out it was some conditional mishap on the loops. Just added an if sttement after all the loops had ended if the second to the last element is non existent and if it is just that second to the last element! Also I made changes to the descending order mode of the function as well since those changes. I discovered that
there is also conditional shit going to happen there so I changed it as well!