FIXED URL ENCODING SHITS WHEN SUBMITTING THE ANSWERS The server can now convert the URI shits to normal string since the stringify function of the JSON class from javascript makes it that way. I first thought of making the shits all replace functions which would be so fucking bad.
I did some improvements when the client can't fetch the core js files so that the pop up would not pop up everytime. I decided it to make the client or the browser load up until 5 times or more so that the user experience is not
that really bad.

So yeah that's it.

I know I should've done that loading improvement on another branch but what's the point?