First changes and improvement on the Admin panel
I'm saying that this is just the first change. I'm not sure if my changes would come up in such high intuitiveness for Lois. I just don't want to make her suffer so that's why care about this kinda useless portion of development. But still it's necessary for this is the fucking site where we would just bang our heads into.