THE REAL URL ENCODING FIX SHIT NOT THE FIX FROM commit f030803e9d265ba2d1be1c61a563a467a509e7
I don't know if that commit hash is accurate. Either way, I found the real culprit behind that issue. Good thing I did a test with my pal of mine, Nico to see if things work out in orderly manner. It was great so far until the answers arrived in the backend. Not all of the fucking
percent escape where decoded and there were still some of it left. I was confused of how that would happen since I made a change in my CreateAnswer class inside the views.py.

I thought the reason why the answers would get escaped was due to the fucking shit being stringified through the JSON.stringify() and I just didn't bother since I have no way of putting it. But I just found out that the reason was the form.serialize I put it. I forgot about that part
and removed it in the equation. That was the reason why the fucking answers would get percent escaped because it was "serialized". That was kinda a gotcha  moment. When I found it, it was like, "ohh, interesting..."

I then removed the supposed "fix" from that URI ENCODING FIX commit and that's it I guess. Everything is now fine after that. The workload is now transferred into the client. Great!